backend/fastapi/pyproject.toml
[tool.poetry]
name = "fastapi-lambda-demo"
version = "0.1.0"
description = "FastAPI Lambda Demo Application"
authors = ["gadname <gakugaku1114@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
mangum = "^0.19.0"
pydantic = "^2.10.6"
sqlalchemy = "^2.0.38"
uvicorn = "^0.27.0"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
ulid-py = "^1.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
httpx = "^0.25.0"
pytest-asyncio = "^0.21.1" # Or a more recent compatible version

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 